Fix to filepaths generated on Windows
Currently, images with file paths generated on Windows may encounter issues when opened on Linux due to differences in path slashes ('/' and '/'). There isn't any problem with opening images with different slashes in the file path on Windows; however, on Linux, the images may not load unless you manually correct the slashes in the annotations.