Replace broken LFW download URLs with working mirrors
The University of Massachusetts server hosting the Labeled Faces in the Wild
dataset is currently unreachable, causing download failures. This updates the
download URLs to use figshare-hosted mirrors that provide identical dataset
files.

Changes:
- Update _VIDEOS_DOWNLOAD_LINK from vis-www.cs.umass.edu to figshare
- Update _TEST_DOWNLOAD_LINK from vis-www.cs.umass.edu to figshare
- Update _TRAIN_DOWNLOAD_LINK from vis-www.cs.umass.edu to figshare

The figshare mirrors are maintained as part of scikit-learn's data hosting
infrastructure and provide stable access to the complete lfw-funneled dataset.