Don't re-queue failed ops to avoid race conditions
Re-queueing leaves `self._sample_ops` and `self._frame_ops` in the same
state that develop does, but because another async save might have been
queued on the executor, re-queueing could cause ops to be written twice.
Unfortunately, those ops are not idempotent, so the DB could end up with
duplicate samples or frames. It's only incidental that develop does not
clear the lists of ops when `bulk_write` fails, and we're not aware of
anything that is done with data that failed to be written (if so, that
would also have non-idempotency problems). Thus, it should be okay to
drop failed samples. (We could write them to a different instance
attribute, but we don't have anything to do with that data.)